pyproject.toml
[project]
name = "unit-converter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.4",
    "streamlit>=1.42.2",
]
